requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
importlib-resources== 6.1.0
kiwisolver==1.4.5
matplotlib==3.8.0
msvc-runtime==14.34.31931
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
pip==23.3.1
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5=5.15.2
pytesseract==0.3.10
python-dateutil==2.8.2
pywin32==306
setuptools==68.2.2
six==1.16.0
wheel==0.41.2
zipp==3.17.0
